+++ components/automate-ui/src/app/modules/infra-proxy/policy-groups/policy-groups.component.scss
@@ -44,3 +44,7 @@ chef-loading-spinner {
     margin-top: 4px;
   }
 }
+
+.policy-group-list-paging {
+  margin: 0 0 35px;
+}
