+++ components/automate-ui/src/app/modules/infra-proxy/policy-groups/policy-groups.component.ts
@@ -35,8 +35,10 @@ export class PolicyGroupsComponent implements OnInit, OnDestroy {
   public authFailure = false;
   public searching = false;
   public searchValue = '';
-  public searchArr;
   public groupList = [];
+  public pageOfItems: Array<any>;
+  public searchFlag: boolean;
+  public policyGroups;
 
   constructor(
     private store: Store<NgrxStateAtom>,
@@ -57,6 +59,7 @@ export class PolicyGroupsComponent implements OnInit, OnDestroy {
         this.policyFiles = policyFilesState;
         this.groupList = this.policyFiles;
         this.policyGroupsListLoading = false;
+        this.filterDataGroupWise();
       } else if (getPolicyFilesSt === EntityStatus.loadingFailure) {
         this.policyGroupsListLoading = false;
         this.authFailure = true;
@@ -81,20 +84,53 @@ export class PolicyGroupsComponent implements OnInit, OnDestroy {
     this.store.dispatch(new GetPolicyGroups(payload));
   }
 
+  filterDataGroupWise() {
+    const key = 'policy_group';
+    this.policyGroups = [];
+    if (this.groupList.length > 0) {
+      this.groupList.forEach((x) => {
+        // Checking if there is any object in this.policyGroupsList
+        // which contains the key value
+        if (this.policyGroups.some((val) => val[key] === x[key])) {
+          // If yes! then increase the occurrence by 1
+          this.policyGroups.forEach((k) => {
+            if (k[key] === x[key]) {
+              k['occurrence']++;
+            }
+          });
+        } else {
+          // If not! Then create a new object initialize
+          // it with the present iteration key's value and set the occurrence to 1
+          const a = {};
+          a[key] = x[key];
+          a['occurrence'] = 1;
+          this.policyGroups.push(a);
+        }
+      });
+    } else {
+      this.policyGroups = [];
+    }
+  }
+
   searchPolicyFiles(searchText: string): void {
     this.searching = true;
     this.searchValue = searchText;
-    if (!this.policyFiles || !searchText) {
-      this.groupList = this.policyFiles;
+    if (!this.policyGroups || !searchText) {
+      this.groupList = this.policyGroups;
+      this.searchFlag = false;
     } else {
-      const list = this.policyFiles.filter((key) => {
+      const list = this.policyGroups.filter((key) => {
+        this.searchFlag = true;
         if (key) {
           return key.policy_group.includes(this.searchValue);
         }
       });
       this.groupList = list;
     }
-    this.searchArr = this.groupList;
     this.searching = false;
+  }
+
+  onChangePage(pageOfItems: Array<any>) {
+    this.pageOfItems = pageOfItems;
   }
 }
